Cleaned up index.js by creating event handler and rewriting command handler
